[PSym] refactoring
