+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/Scheduler.java
@@ -12,8 +12,8 @@ import psym.runtime.machine.Monitor;
 import psym.runtime.machine.State;
 import psym.runtime.machine.events.Event;
 import psym.runtime.machine.events.Message;
-import psym.runtime.scheduler.symmetry.SymmetryMode;
-import psym.runtime.scheduler.symmetry.SymmetryTracker;
+import psym.runtime.scheduler.search.symmetry.SymmetryMode;
+import psym.runtime.scheduler.search.symmetry.SymmetryTracker;
 import psym.runtime.statistics.SearchStats;
 import psym.utils.Assert;
 import psym.utils.random.NondetUtil;
