+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/search/symmetry/SymmetryMode.java
@@ -1,0 +1,6 @@
+package psym.runtime.scheduler.search.symmetry;
+
+public enum SymmetryMode {
+  None,
+  Full
+}
