+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/search/taskorchestration/TaskOrchestrator.java
@@ -1,7 +1,6 @@
-package psym.runtime.scheduler.explicit.taskorchestration;
+package psym.runtime.scheduler.search.taskorchestration;
 
 import java.io.Serializable;
-import psym.runtime.scheduler.explicit.BacktrackTask;
 
 public interface TaskOrchestrator extends Serializable {
     void addPriority(BacktrackTask task);
