+++ Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/search/taskorchestration/TaskOrchestratorRandom.java
@@ -1,7 +1,7 @@
-package psym.runtime.scheduler.explicit.taskorchestration;
+package psym.runtime.scheduler.search.taskorchestration;
 
 import java.util.*;
-import psym.runtime.scheduler.explicit.BacktrackTask;
+
 import psym.utils.random.RandomNumberGenerator;
 
 public class TaskOrchestratorRandom implements TaskOrchestrator {
