--- Src/PRuntimes/PSymRuntime/src/main/java/psym/runtime/scheduler/symmetry/SymmetryMode.java
@@ -1,6 +1,0 @@
-package psym.runtime.scheduler.symmetry;
-
-public enum SymmetryMode {
-  None,
-  Full
-}
